Python Refactor: Rename NerApproach module
