+++ python/sparknlp/annotator/ner/ner_dl.py
@@ -16,7 +16,7 @@
 import sys
 
 from sparknlp.common import *
-from sparknlp.annotator.ner import NerApproach
+from sparknlp.annotator.ner.ner_approach import NerApproach
 
 
 class NerDLApproach(AnnotatorApproach, NerApproach):
